Added Layer type.

+++ src/libs/rand.cpp
@@ -4,7 +4,7 @@ void seed_rand(void) {
 	srand((unsigned int) time(NULL));
 }
 
-double rand_weight(void) {
-    double weight = (double)rand() / RAND_MAX;
+float rand_weight(void) {
+    float weight = (float)rand() / RAND_MAX;
     return (weight * 2) - 1;
 }
